use size_t

+++ src/filters.d
@@ -7,7 +7,7 @@ enum Growable { no, yes }
 enum Copyable { no, yes }
 
 enum isDenseSetFilterable(E) = (is(typeof(cast(size_t)E.init)) && // is castable to size_t
-                                cast(uint)E.max <= uint.max);      // and small enough
+                                cast(size_t)E.max <= uint.max);      // and small enough
 
 /** Store presence of elements of type `E` in a set in the range `0 .. length`.
     Can be seen as a generalization of `std.typecons.BitFlags` to integer types.
